pyproject.toml
[project]
name = "certificate-summary"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles>=25.1.0",
    "alembic>=1.18.1",
    "asyncpg>=0.31.0",
    "black>=26.1.0",
    "camoufox>=0.4.11",
    "fastapi[standard]>=0.128.0",
    "ldap3>=2.9.1",
    "openpyxl>=3.1.5",
    "psycopg>=3.3.2",
    "pydantic-settings>=2.12.0",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "uvicorn>=0.40.0",
]
